refactor: rename and reorganize markdown and tooltip components
- Renamed `markdownComponents.tsx` to `markdown-components.tsx` and updated references accordingly.
- Renamed `tooltip.tsx` to `suggestions-tooltip.tsx` and updated references in the message suggestions module.
- Removed the old tooltip implementation to streamline the codebase.
