+++ cli/src/registry/message-suggestions/message-suggestions.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { MessageGenerationStage } from "@/components/ui/message-generation-stage";
-import { Tooltip, TooltipProvider } from "@/components/ui/tooltip";
+import { Tooltip, TooltipProvider } from "@/components/ui/suggestions-tooltip";
 import { cn } from "@/lib/utils";
 import type { Suggestion, TamboThread } from "@tambo-ai/react";
 import { useTambo, useTamboSuggestions } from "@tambo-ai/react";
