+++ cli/src/registry/message-suggestions/suggestions-tooltip.tsx
